refactor(codemod): rename `queryKey` replacer and error, because now the key name can be parameterized
