+++ codemods/v4/use_mutation/use-mutation.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const createUtilsObject = require('../utils')
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const createQueryKeyReplacer = require('../utils/replacers/query-key-replacer')
+const createKeyReplacer = require('../utils/replacers/key-replacer')
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const hookCallTransformer = require('../utils/transformers/hook-call-transformer')
 
@@ -10,7 +10,7 @@ module.exports = (file, api) => {
   const root = jscodeshift(file.source)
 
   const utils = createUtilsObject({ root, jscodeshift })
-  const replacer = createQueryKeyReplacer({
+  const replacer = createKeyReplacer({
     jscodeshift,
     root,
     keyName: 'mutationKey',
